refactor: refactor component folder structure
