+++ src/pages/components/Header.tsx
@@ -1,0 +1,9 @@
+import { IoIosArrowBack } from "react-icons/io";
+
+export default function Header() {
+  return (
+    <header>
+      <IoIosArrowBack />
+    </header>
+  );
+}
